Improve display when JavaScript is unavailable.
Rather than have the no-transforms class added to the body if the
browser doesn't pass the transform support tests, this commit
reverses this so that the no-transforms class is there by default, and
is then removed if the transform support tests pass.

This way round, it means that if the JavaScript does not load for any
reason (e.g. bad connectivity) or is otherwise unavailable, the
content is still visible and able to be read. This raises "all non-
JavaScript browsers" from the currently described 5th tier to 4th tier.

It also tweaks the CSS so that fragments and speaker notes are visibly
displayed in this situation.